feat(image-upload): refactor image upload handling and replace shared component with dedicated implementation
