+++ apps/web/src/components/tenant/image-upload-area.tsx
